--- openml-lightgbm-provider/pom.xml
@@ -1,195 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ The copyright of this file belongs to Feedzai. The file cannot be
-  ~ reproduced in whole or in part, stored in a retrieval system,
-  ~ transmitted in any form, or by any means electronic, mechanical,
-  ~ photocopying, or otherwise, without the prior permission of the owner.
-  ~
-  ~ (c) 2020 Feedzai, Strictly Confidential
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>openml-java</artifactId>
-        <groupId>com.feedzai</groupId>
-        <version>0.0.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>openml-lightgbm-provider</artifactId>
-
-    <name>Feedzai OpenML LGBM Provider</name>
-    <description>Contains an example for the implementation of a LGBM Machine Learning Model and CLassifier
-    </description>
-
-    <properties>
-        <openml.version>1.0.3</openml.version>
-        <lightgbmlib.version>2.3.150</lightgbmlib.version>
-    </properties>
-
-
-    <dependencies>
-        <dependency>
-            <groupId>com.feedzai</groupId>
-            <artifactId>openml-api</artifactId>
-            <version>${openml.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.feedzai</groupId>
-            <artifactId>openml-utils</artifactId>
-            <version>${openml.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.microsoft.ml.lightgbm</groupId>
-            <artifactId>lightgbmlib</artifactId>
-            <version>${lightgbmlib.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.google.auto.service</groupId>
-            <artifactId>auto-service</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-        </dependency>
-
-        <!-- For testing only -->
-        <dependency>
-            <groupId>ch.qos.logback</groupId>
-            <artifactId>logback-core</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.feedzai</groupId>
-            <artifactId>openml-utils</artifactId>
-            <type>test-jar</type>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.assertj</groupId>
-            <artifactId>assertj-core</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-csv</artifactId>
-            <version>1.8</version>
-            <scope>test</scope>
-        </dependency>
-
-    </dependencies>
-
-    <build>
-
-        <resources>
-            <resource>
-                <directory>${project.basedir}/target/resources</directory>
-            </resource>
-        </resources>
-
-        <plugins>
-            <plugin>
-                <artifactId>maven-assembly-plugin</artifactId>
-                <configuration>
-                    <descriptorRefs>
-                        <descriptorRef>jar-with-dependencies</descriptorRef>
-                    </descriptorRefs>
-                    <appendAssemblyId>false</appendAssemblyId>
-                </configuration>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>single</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-
-            <!-- set locale otherwise the LightGBM core implementation spits wrong results -->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <configuration>
-                    <environmentVariables>
-                        <LC_ALL>C</LC_ALL>
-                    </environmentVariables>
-                </configuration>
-            </plugin>
-
-            <!-- Install the local copy of the custom build of LightGBM to Maven -->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-install-plugin</artifactId>
-                <version>3.0.0-M1</version>
-                <executions>
-                    <execution>
-                        <id>install-custom-lightgbmlib_build</id>
-                        <phase>initialize</phase>
-                        <goals>
-                            <goal>install-file</goal>
-                        </goals>
-                        <configuration>
-                            <file>${project.basedir}/lightgbmlib_build/lightgbmlib.jar</file>
-                            <pomFile>${project.basedir}/lightgbmlib_build/pom.xml</pomFile>
-                            <generatePom>false</generatePom>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <!-- Copy the shared libraries to the resources folder -->
-            <plugin>
-                <artifactId>maven-resources-plugin</artifactId>
-                <version>3.1.0</version>
-                <executions>
-                    <execution>
-                        <id>copy-shared-libs-resources</id>
-                        <phase>generate-resources</phase>
-                        <goals>
-                            <goal>copy-resources</goal>
-                        </goals>
-                        <configuration>
-                            <outputDirectory>${project.basedir}/target/resources</outputDirectory>
-                            <resources>
-                                <resource>
-                                    <directory>lightgbmlib_build</directory>
-                                    <include>*.so</include>
-                                    <include>*.so.1.0.0</include>
-                                </resource>
-                            </resources>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>copy-licences</id>
-                        <phase>generate-resources</phase>
-                        <goals>
-                          <goal>copy-resources</goal>
-                        </goals>
-                        <configuration>
-                            <outputDirectory>${project.basedir}/target/resources</outputDirectory>
-                            <resources>
-                                <resource>
-                                    <directory>${maven.multiModuleProjectDirectory}/LICENCES</directory>
-                                </resource>
-                            </resources>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-</project>
