+++ pom.xml
@@ -29,7 +29,7 @@
         <module>openml-datarobot</module>
         <module>openml-h2o</module>
         <module>openml-java-utils</module>
-        <module>openml-lightgbm-provider</module>
+        <module>openml-lightgbm</module>
     </modules>
 
     <name>Java OpenML Main</name>
@@ -76,6 +76,7 @@
         <jackson.version>2.6.7</jackson.version>
         <jackson-databind.version>2.6.7</jackson-databind.version>
         <commons-io.version>2.4</commons-io.version>
+        <apache.version>1.8</apache.version>
     </properties>
 
     <dependencyManagement>
@@ -153,6 +154,12 @@
                 <version>${assertj.version}</version>
                 <scope>test</scope>
             </dependency>
+            <dependency>
+                <groupId>org.apache.commons</groupId>
+                <artifactId>commons-csv</artifactId>
+                <version>${apache.version}</version>
+                <scope>test</scope>
+            </dependency>
         </dependencies>
     </dependencyManagement>
 
@@ -187,6 +194,16 @@
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-shade-plugin</artifactId>
                     <version>3.2.0</version>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-install-plugin</artifactId>
+                    <version>3.0.0-M1</version>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-resources-plugin</artifactId>
+                    <version>3.1.0</version>
                 </plugin>
             </plugins>
         </pluginManagement>
